fix changes

+++ activity/src/main/java/com/kairos/persistence/model/time_bank/DailyTimeBankEntry.java
@@ -39,7 +39,7 @@ public class DailyTimeBankEntry extends MongoBaseEntity{
     //It is the sum of scheduledMinutesOfTimeBank + ctaBonusMinutesOfTimeBank
     private int plannedMinutesOfTimebank;
     private Map<LocalDate,Integer> publishedBalances;
-    private int protectedDaysOffHours;
+    private long protectedDaysOffHours;
     private DailyTimeBankEntry draftDailyTimeBankEntry;
